+++ src/index.ts
@@ -8,6 +8,7 @@ This code is licensed under the BSD-3 license found in the LICENSE file in the r
 export { blueTheme as blue } from "./blueTheme";
 export { blueDarkTheme as blueDark } from "./blueDarkTheme";
 import { $DeepPartial } from "@callstack/react-theme-provider";
+import { ThemeOpacity, ThemeOverrides } from "./shared";
 
 // This will still need to be added to an end-user project if they want to access these properties without typescript yelling at them.
 declare global {
@@ -48,6 +49,7 @@ declare global {
                 disabledBackground: string;
             };
             overrides: $DeepPartial<ThemeOverrides>;
+            opacity: Partial<ThemeOpacity>;
         }
         // eslint-disable-next-line @typescript-eslint/consistent-type-definitions
         interface ThemeFonts {
@@ -55,42 +57,3 @@ declare global {
         }
     }
 }
-
-export type ThemeOverrides = {
-    avatar: {
-        background: string;
-    };
-
-    bottomNavigation: {
-        inactive: string;
-    };
-    button: {
-        contained: {
-            text: {
-                disabled: string;
-            };
-        };
-    };
-    chip: {
-        outlined: {
-            background: string;
-        };
-        flat: {
-            background: string;
-        };
-    };
-    snackbar: {
-        accent: string;
-    };
-    toggleButtonGroup: {
-        checked: {
-            background: string;
-        };
-    };
-};
-
-export type ThemeOpacity = {
-    disabled: number;
-    divider: number;
-    disabledBackground: number;
-};
